AmigoSecreto/requirements.txt
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
Flask-Bcrypt==1.0.1
Flask-Login>=0.6.2
email-validator==1.3.1
flasgger==0.9.5
python-dotenv==1.0.0
sib-api-v3-sdk==7.6.0
mysql-connector-python==8.0.33
pymysql==1.0.3
gunicorn==20.1.0
Werkzeug==2.0.3
flask-migrate==3.0.1
flask-socketio==5.4.1
SQLAlchemy==1.4.49
requests==2.31.0